Development

+++ ext/python/tttr/tttr_extension.py
@@ -30,11 +30,11 @@ def __len__(self):
 
 def __getitem__(self, key):
     if isinstance(key, slice):
-        sel = np.arange(*key.indices(self.get_n_valid_events()), dtype=np.int64)
+        sel = np.arange(*key.indices(self.get_n_valid_events()), dtype=np.int32)
     elif isinstance(key, np.ndarray):
-        sel = key.astype(np.int64)
+        sel = key.astype(np.int32)
     else:
-        sel = np.array([key], dtype=np.int64)
+        sel = np.array([key], dtype=np.int32)
     return tttrlib.TTTR(self, sel)
